requirements.txt
mkdocs>=1.6.0
mkdocs-material>=9.5.0
mkdocs-minify-plugin>=0.8.0
requests>=2.31.0
feedparser>=6.0.0
jinja2>=3.1.0
python-dateutil>=2.8.0
matplotlib>=3.8.0
pyyaml>=6.0
